Feat/add authentication

+++ main/views.py
@@ -1,4 +1,10 @@
+import datetime
+
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth import login, logout
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from django.contrib.auth.decorators import login_required  
+from django.core.exceptions import PermissionDenied        
 from django.contrib import messages
 from django.core import serializers
 from django.http import HttpResponse
@@ -7,10 +13,50 @@ from main.models import Experience
 from main.models import Skill
 from .forms import ExperienceForm, SkillForm
 
-''' ===============================
+''' =============================
+    Functions for Authentication
+    ============================= '''
+def register(request):
+    form = UserCreationForm(request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        form.save()
+        messages.success(request, "Akun berhasil dibuat. Silakan login.")
+        return redirect("main:login")
+
+    context = {
+        "name": "Nanta",
+        "form": form,
+    }
+    return render(request, "register.html", context)
+
+def login_user(request):
+    form = AuthenticationForm(request, data=request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        user = form.get_user()
+        login(request, user)
+        response = redirect("main:show_main")
+        response.set_cookie('last_login', datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
+        return response
+
+    context = {
+        "name": "Nanta",
+        "form": form,
+    }
+    return render(request, "login.html", context)
+
+def logout_user(request):
+    logout(request)
+    response = redirect("main:show_main")
+    response.delete_cookie('last_login')
+    return response;
+
+''' =============================== 
     Functions for Main Landing Page
     =============================== '''
 def show_main(request):
+    last_login = request.COOKIES.get('last_login', 'Belum ada sesi login / Cookie tidak ditemukan')
     context = {
         "name": "Nanta",
         "full_name" : "Muhammad Raihananta Adzaky Yuwono",
@@ -19,6 +65,7 @@ def show_main(request):
         "bio": (
             "CS student at the University of Indonesia. I'm very eager to grow my skills in programming, software development and cybersecurity."
         ),
+        "last_login": last_login,
     }
     return render(request, "index.html", context)
 
@@ -42,7 +89,11 @@ def show_experience(request):
     }
     return render(request, "experience.html", context)
 
+@login_required(login_url="/login/")
 def create_experience(request):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+    
     form = ExperienceForm(request.POST or None)
 
     if request.method == "POST" and form.is_valid():
@@ -57,7 +108,11 @@ def create_experience(request):
     }
     return render(request, "experience_form.html", context)
 
+@login_required(login_url="/login/")
 def edit_experience(request, experience_id):
+    if not request.user.is_superuser:
+            raise PermissionDenied
+    
     experience = get_object_or_404(Experience, id=experience_id)
     form = ExperienceForm(request.POST or None, instance=experience)
 
@@ -82,16 +137,34 @@ def get_experience_json(request):
     if title_query:
         experience_list = experience_list.filter(title__icontains=title_query)
 
-    experience_json = serializers.serialize("json", experience_list)
+    experience_json = serializers.serialize("json", experience_list, use_natural_foreign_keys=True)
     return HttpResponse(experience_json, content_type="application/json")
 
+@login_required(login_url="/login/")
 def delete_experience(request, experience_id):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+    
     experience = get_object_or_404(Experience, pk=experience_id)
 
     if request.method == "POST":
         experience.delete()
         messages.success(request, "Pengalaman berhasil dihapus!")
         return redirect("main:show_experience")
+
+    return redirect("main:show_experience")
+
+@login_required(login_url="/login/")
+def toggle_star(request, experience_id):
+    experience = get_object_or_404(Experience, pk=experience_id)
+
+    if request.method == "POST":
+        # Kalau akun ini sudah pernah memberi star, batalkan star-nya.
+        # Kalau belum, tambahkan star
+        if request.user in experience.starred_by.all():
+            experience.starred_by.remove(request.user)
+        else:
+            experience.starred_by.add(request.user)
 
     return redirect("main:show_experience")
